requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
langchain_experimental
sse_starlette
bs4
pypdf
chromadb
faiss-cpu